+++ src/Candoumbe.Types.Strings/StringSegmentLinkedList.cs
@@ -1,9 +1,8 @@
 using System;
+using System.Buffers;
 using System.Collections;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
-using System.Text;
 using Candoumbe.MiscUtilities.Comparers;
 using Microsoft.Extensions.Primitives;
 
@@ -20,13 +19,14 @@ public class StringSegmentLinkedList : IEnumerable<ReadOnlyMemory<char>>, IEquat
     private StringSegmentNode _head;
     private StringSegmentNode _tail;
 
-    private static StringSegmentNode EmptyNode => new([]);
-
     /// <summary>
     /// Builds a new instance of <see cref="StringSegmentLinkedList"/> that is empty.
     /// </summary>
-    public StringSegmentLinkedList() : this(ReadOnlySpan<char>.Empty)
+    public StringSegmentLinkedList()
     {
+        _head = null;
+        _tail = null;
+        Count = 0;
     }
 
     /// <summary>
@@ -35,9 +35,71 @@ public class StringSegmentLinkedList : IEnumerable<ReadOnlyMemory<char>>, IEquat
     /// <param name="head">The value of the head</param>
     public StringSegmentLinkedList(ReadOnlySpan<char> head)
     {
-        (_head, Count) = head.IsEmpty
-                             ? (EmptyNode, 1)
-                             : (new StringSegmentNode(head), 1);
+        if (head.IsEmpty)
+        {
+            _head = _tail = null;
+            Count = 0;
+        }
+        else
+        {
+            _head = _tail = new StringSegmentNode(head);
+            Count = 1;
+        }
+    }
+
+    /// <summary>
+    /// Builds a new instance of <see cref="StringSegmentLinkedList"/>.
+    /// </summary>
+    /// <param name="head">The value of the head</param>
+    public StringSegmentLinkedList(string head)
+    {
+        if (string.IsNullOrEmpty(head))
+        {
+            _head = _tail = null;
+            Count = 0;
+        }
+        else
+        {
+            _head = _tail = new StringSegmentNode(head);
+            Count = 1;
+        }
+    }
+
+    /// <summary>
+    /// Builds a new instance of <see cref="StringSegmentLinkedList"/>.
+    /// </summary>
+    /// <param name="head">The value of the head</param>
+    public StringSegmentLinkedList(ReadOnlyMemory<char> head)
+    {
+        if (head.IsEmpty)
+        {
+            _head = _tail = null;
+            Count = 0;
+        }
+        else
+        {
+            _head = _tail = new StringSegmentNode(head);
+            Count = 1;
+        }
+    }
+
+    /// <summary>
+    /// Builds a new instance of <see cref="StringSegmentLinkedList"/>.
+    /// </summary>
+    /// <param name="head">The head StringSegment</param>
+    public StringSegmentLinkedList(StringSegment head)
+    {
+        if (!head.HasValue || head.Length == 0)
+        {
+            _head = _tail = null;
+            Count = 0;
+        }
+        else
+        {
+            ReadOnlyMemory<char> mem = head.Buffer?.AsMemory(head.Offset, head.Length) ?? ReadOnlyMemory<char>.Empty;
+            _head = _tail = new StringSegmentNode(mem);
+            Count = 1;
+        }
     }
 
     /// <summary>
@@ -45,14 +107,15 @@ public class StringSegmentLinkedList : IEnumerable<ReadOnlyMemory<char>>, IEquat
     /// </summary>
     /// <param name="head">the value of the head</param>
     /// <param name="others">Additional elements to append to the list.</param>
-    public StringSegmentLinkedList(ReadOnlySpan<char> head, params ReadOnlySpan<StringSegment> others)
+    public StringSegmentLinkedList(string head, params string[] others)
+        : this(head)
     {
-        _head = new StringSegmentNode(head);
-        Count = 1;
-
-        foreach (ReadOnlySpan<char> next in others)
+        if (others is { Length: > 0 })
         {
-            Append(next);
+            foreach (string s in others)
+            {
+                Append(s);
+            }
         }
     }
 
@@ -60,6 +123,7 @@ public class StringSegmentLinkedList : IEnumerable<ReadOnlyMemory<char>>, IEquat
     /// Appends <paramref name="value"/> to the end of the list.
     /// </summary>
     /// <param name="value">The value to append</param>
+    /// <returns>The current instance with <paramref name="value"/> at the tail.</returns>
     /// <remarks>The current instance remains untouched if <paramref name="value"/> is empty.</remarks>
     public StringSegmentLinkedList Append(ReadOnlySpan<char> value)
     {
@@ -72,24 +136,77 @@ public class StringSegmentLinkedList : IEnumerable<ReadOnlyMemory<char>>, IEquat
         return this;
     }
 
+    /// <summary>
+    /// Appends <paramref name="value"/> to the end of the list.
+    /// </summary>
+    /// <param name="value">The value to append</param>
+    /// <returns>The current instance with <paramref name="value"/> at the tail.</returns>
+    /// <remarks>The current instance remains untouched if <paramref name="value"/> is empty.</remarks>
+    public StringSegmentLinkedList Append(StringSegment value)
+    {
+        if (value is { HasValue: true, Length: > 0 })
+        {
+            ReadOnlyMemory<char> mem = value.Buffer?.AsMemory(value.Offset, value.Length) ?? ReadOnlyMemory<char>.Empty;
+            if (!mem.IsEmpty)
+            {
+                AppendInternal(new StringSegmentNode(mem));
+            }
+        }
+
+        return this;
+    }
+
+    /// <summary>
+    /// Appends <paramref name="value"/> to the end of the list.
+    /// </summary>
+    /// <param name="value">The value to append</param>
+    /// <returns>The current instance with <paramref name="value"/> at the tail.</returns>
+    /// <remarks>The current instance remains untouched if <paramref name="value"/> is empty.</remarks>
+    public StringSegmentLinkedList Append(string value)
+    {
+        if (!string.IsNullOrEmpty(value))
+        {
+            AppendInternal(new StringSegmentNode(value));
+        }
+
+        return this;
+    }
+
+    /// <summary>
+    /// Appends <paramref name="value"/> to the end of the list.
+    /// </summary>
+    /// <param name="value">The value to append</param>
+    /// <returns>The current instance with <paramref name="value"/> at the tail.</returns>
+    /// <remarks>The current instance remains untouched if <paramref name="value"/> is empty.</remarks>
+    public StringSegmentLinkedList Append(ReadOnlyMemory<char> value)
+    {
+        if (!value.IsEmpty)
+        {
+            AppendInternal(new StringSegmentNode(value));
+        }
+
+        return this;
+    }
+
+    /// <summary>
+    /// Appends <paramref name="value"/> to the end of the list.
+    /// </summary>
+    /// <param name="value">The value to append</param>
+    /// <returns>The current instance with <paramref name="value"/> at the tail.</returns>
+    /// <remarks>The current instance remains untouched if <paramref name="value"/> is empty.</remarks>
+    public StringSegmentLinkedList Append(char[] value) => value is { Length: > 0 }
+                                                               ? Append((ReadOnlyMemory<char>)value)
+                                                               : this;
+
     private void AppendInternal(StringSegmentNode newNode)
     {
-        if (_tail is null)
+        if (_head is null)
         {
-            _tail = newNode;
-            if (_head.Equals(EmptyNode))
-            {
-                _head = newNode;
-            }
-            else
-            {
-                _head.Next = _tail;
-            }
+            _head = _tail = newNode;
         }
         else
         {
-            _head.Next ??= _tail;
-            _tail.Next = newNode;
+            _tail!.Next = newNode;
             _tail = newNode;
         }
 
@@ -102,6 +219,7 @@ public class StringSegmentLinkedList : IEnumerable<ReadOnlyMemory<char>>, IEquat
     /// <param name="index">0-based index where <paramref name="value"/> will be inserted</param>
     /// <param name="value">The value of the node to insert.</param>
     /// <exception cref="ArgumentOutOfRangeException">if <paramref name="index"/> is &lt; 0. or &gt; <see cref="Count"/>.</exception>
+    /// <remarks>The current instance remains untouched if <paramref name="value"/> is empty.</remarks>
     public void InsertAt(int index, ReadOnlySpan<char> value)
     {
         if (index < 0 || index > Count)
@@ -116,27 +234,104 @@ public class StringSegmentLinkedList : IEnumerable<ReadOnlyMemory<char>>, IEquat
         }
     }
 
+    /// <summary>
+    /// Inserts a new link containing <paramref name="value"/> at the given <paramref name="index"/>.
+    /// </summary>
+    /// <param name="index">0-based index where <paramref name="value"/> will be inserted</param>
+    /// <param name="value">The value of the node to insert.</param>
+    /// <remarks>The current instance remains untouched if <paramref name="value"/> is empty.</remarks>
+    /// <exception cref="ArgumentOutOfRangeException">if <paramref name="index"/> is &lt; 0 or &gt; <see cref="Count"/>.</exception>
+    public void InsertAt(int index, StringSegment value)
+    {
+        if (index < 0 || index > Count)
+        {
+            throw new ArgumentOutOfRangeException(nameof(index), "Index is out of range.");
+        }
+
+        if (value is { HasValue: true, Length: > 0 })
+        {
+            ReadOnlyMemory<char> mem = value.Buffer?.AsMemory(value.Offset, value.Length) ?? ReadOnlyMemory<char>.Empty;
+            if (!mem.IsEmpty)
+            {
+                InsertAtInternal(index, new StringSegmentNode(mem));
+            }
+        }
+    }
+
+    /// <summary>
+    /// Inserts a new link containing <paramref name="value"/> at the given <paramref name="index"/>.
+    /// </summary>
+    /// <param name="index">0-based index where <paramref name="value"/> will be inserted</param>
+    /// <param name="value">The value of the node to insert.</param>
+    /// <remarks>The current instance remains untouched if <paramref name="value"/> is empty.</remarks>
+    /// <exception cref="ArgumentOutOfRangeException">if <paramref name="index"/> is &lt; 0 or &gt; <see cref="Count"/>.</exception>
+    public void InsertAt(int index, string value)
+    {
+        if (index < 0 || index > Count)
+        {
+            throw new ArgumentOutOfRangeException(nameof(index), "Index is out of range.");
+        }
+
+        if (!string.IsNullOrEmpty(value))
+        {
+            InsertAtInternal(index, new StringSegmentNode(value));
+        }
+    }
+
+    /// <summary>
+    /// Inserts a new link containing <paramref name="value"/> at the given <paramref name="index"/>.
+    /// </summary>
+    /// <param name="index">0-based index where <paramref name="value"/> will be inserted</param>
+    /// <param name="value">The value of the node to insert.</param>
+    /// <remarks>The current instance remains untouched if <paramref name="value"/> is empty.</remarks>
+    /// <exception cref="ArgumentOutOfRangeException">if <paramref name="index"/> is &lt; 0 or &gt; <see cref="Count"/>.</exception>
+    public void InsertAt(int index, ReadOnlyMemory<char> value)
+    {
+        if (index < 0 || index > Count)
+        {
+            throw new ArgumentOutOfRangeException(nameof(index), "Index is out of range.");
+        }
+
+        if (!value.IsEmpty)
+        {
+            InsertAtInternal(index, new StringSegmentNode(value));
+        }
+    }
+
     private void InsertAtInternal(int index, StringSegmentNode newNode)
     {
-        if (index is 0)
+        if (index == 0)
         {
             newNode.Next = _head;
             _head = newNode;
+            if (_tail is null)
+            {
+                _tail = newNode;
+            }
+        }
+        else if (index == Count)
+        {
+            // insert at end
+            if (_tail is null)
+            {
+                _head = _tail = newNode;
+            }
+            else
+            {
+                _tail.Next = newNode;
+                _tail = newNode;
+            }
         }
         else
         {
-            StringSegmentNode current = _head;
-            StringSegmentNode previous = null;
-
-            for (int i = 0; i < index; i++)
+            StringSegmentNode previous = _head;
+            for (int i = 1; i < index; i++)
             {
-                previous = current;
-                current = current!.Next;
+                previous = previous!.Next;
             }
 
-            previous ??= newNode;
+            newNode.Next = previous!.Next;
             previous.Next = newNode;
-            newNode.Next ??= _tail;
         }
 
         Count++;
@@ -175,6 +370,77 @@ public class StringSegmentLinkedList : IEnumerable<ReadOnlyMemory<char>>, IEquat
     /// </summary>
     public int Count { get; private set; }
 
+    // Configuration/Heuristics
+    public int ChunkSize { get; set; } = 4096;
+    public int MaxSmallSegmentLength { get; set; } = 64;
+    public int MaxNodeCountBeforeCompact { get; set; } = 10_000;
+    public int ReplaceCloneThreshold { get; set; } = 1_000_000;
+
+    /// <summary>
+    /// Compacts the list into larger chunks to reduce node count and improve locality.
+    /// </summary>
+    public StringSegmentLinkedList Compact(int targetChunkSize = 4096)
+    {
+        if (_head is null)
+        {
+            return this;
+        }
+
+        ArrayPool<char> pool = ArrayPool<char>.Shared;
+        char[] buffer = pool.Rent(targetChunkSize);
+        try
+        {
+            int pos = 0;
+            StringSegmentLinkedList compacted = new();
+
+            void Flush()
+            {
+                if (pos == 0)
+                {
+                    return;
+                }
+
+                string chunk = new string(buffer, 0, pos);
+                compacted.Append(chunk);
+                pos = 0;
+            }
+
+            StringSegmentNode node = _head;
+            while (node is not null)
+            {
+                ReadOnlySpan<char> span = node.Value.Span;
+                int remaining = span.Length;
+                int offset = 0;
+
+                while (remaining > 0)
+                {
+                    int toCopy = Math.Min(remaining, targetChunkSize - pos);
+                    span.Slice(offset, toCopy).CopyTo(buffer.AsSpan(pos));
+                    pos += toCopy;
+                    offset += toCopy;
+                    remaining -= toCopy;
+                    if (pos == targetChunkSize)
+                    {
+                        Flush();
+                    }
+                }
+
+                node = node.Next;
+            }
+
+            Flush();
+
+            _head = compacted._head;
+            _tail = compacted._tail;
+            Count = compacted.Count;
+            return this;
+        }
+        finally
+        {
+            pool.Return(buffer);
+        }
+    }
+
     /// <summary>
     /// Computes the total length of the resulting string value
     /// </summary>
@@ -195,32 +461,35 @@ public class StringSegmentLinkedList : IEnumerable<ReadOnlyMemory<char>>, IEquat
     /// <summary>
     /// Computes the <see cref="string"/> value.
     /// </summary>
-    /// <returns></returns>
+    /// <returns>The computed <see cref="string"/>.</returns>
     public string ToStringValue()
     {
-        StringBuilder sb = new();
-        StringSegmentNode current = _head;
-        while (current is not null)
+        int total = GetTotalLength();
+        string result = string.Empty;
+
+        if (total is not 0)
         {
-            ReadOnlyMemory<char> segment = current.Value;
-            sb.Append(segment.Span);
-            current = current.Next;
+            result = string.Create(total, _head, static (dest, head) =>
+                                                 {
+                                                     int offset = 0;
+                                                     StringSegmentNode current = head;
+                                                     while (current is not null)
+                                                     {
+                                                         ReadOnlySpan<char> span = current.Value.Span;
+                                                         span.CopyTo(dest[offset..]);
+                                                         offset += span.Length;
+                                                         current = current.Next;
+                                                     }
+                                                 });
         }
 
-        return sb.ToString();
+        return result;
     }
 
     /// <inheritdoc />
     public IEnumerator<ReadOnlyMemory<char>> GetEnumerator()
     {
-        if (_head.Equals(EmptyNode))
-        {
-            yield break;
-        }
-
-        yield return _head.Value;
-
-        StringSegmentNode current = _head.Next;
+        StringSegmentNode current = _head;
         while (current is not null)
         {
             yield return current.Value;
@@ -283,8 +552,14 @@ public class StringSegmentLinkedList : IEnumerable<ReadOnlyMemory<char>>, IEquat
     /// </returns>
     public StringSegmentLinkedList Replace(Func<char, bool> predicate, ReadOnlySpan<char> replacement)
     {
-        StringSegmentLinkedList replacementList = [];
+        if (predicate is null)
+        {
+            throw new ArgumentNullException(nameof(predicate));
+        }
+
+        StringSegmentLinkedList replacementList = new();
         StringSegmentNode current = _head;
+        bool anyReplacement = false;
 
         while (current is not null)
         {
@@ -293,11 +568,11 @@ public class StringSegmentLinkedList : IEnumerable<ReadOnlyMemory<char>>, IEquat
 
             if (indexOfOldChar >= 0)
             {
-                ReadOnlySpan<char> valueToKeep;
+                anyReplacement = true;
+
                 if (indexOfOldChar > 0)
                 {
-                    valueToKeep = current.Value[..indexOfOldChar].Span;
-                    replacementList = replacementList.Append(valueToKeep);
+                    replacementList = replacementList.Append(current.Value[..indexOfOldChar]);
                 }
 
                 replacementList = replacementList.Append(replacement);
@@ -307,32 +582,27 @@ public class StringSegmentLinkedList : IEnumerable<ReadOnlyMemory<char>>, IEquat
                 {
                     if (index < occurrence)
                     {
-                        valueToKeep = current.Value.Span.Slice(index, occurrence - index);
-                        replacementList = replacementList.Append(valueToKeep);
+                        replacementList = replacementList.Append(current.Value.Slice(index, occurrence - index));
                     }
 
                     replacementList = replacementList.Append(replacement);
-
-                    // move the cursor right after the current occurrence
                     index = occurrence + 1;
                 }
 
-                // we did all substitutions, but we did not reach the end of the original input
-                // => copy all remaining original chars starting at the "index" position
                 if (index < current.Value.Length)
                 {
-                    replacementList = replacementList.Append(current.Value[index..].Span);
+                    replacementList = replacementList.Append(current.Value[index..]);
                 }
             }
             else
             {
-                replacementList = replacementList.Append(current.Value.Span);
+                replacementList = replacementList.Append(current.Value);
             }
 
             current = current.Next;
         }
 
-        return replacementList;
+        return anyReplacement ? replacementList : this;
     }
 
     /// <summary>
@@ -357,10 +627,21 @@ public class StringSegmentLinkedList : IEnumerable<ReadOnlyMemory<char>>, IEquat
     /// // `list` will have 'a' replaced with "alpha" and 'b' with "beta"
     /// </code>
     /// </example>
-public StringSegmentLinkedList Replace(Func<char, bool> predicate, IReadOnlyDictionary<char, ReadOnlyMemory<char>> replacements)
+    public StringSegmentLinkedList Replace(Func<char, bool> predicate, IReadOnlyDictionary<char, ReadOnlyMemory<char>> replacements)
     {
-        StringSegmentLinkedList replacementList = [];
+        if (predicate is null)
+        {
+            throw new ArgumentNullException(nameof(predicate));
+        }
+
+        if (replacements is null)
+        {
+            throw new ArgumentNullException(nameof(replacements));
+        }
+
+        StringSegmentLinkedList replacementList = new();
         StringSegmentNode current = _head;
+        bool anyReplacement = false;
 
         while (current is not null)
         {
@@ -369,22 +650,19 @@ public StringSegmentLinkedList Replace(Func<char, bool> predicate, IReadOnlyDict
 
             if (indexOfOldChar >= 0)
             {
+                anyReplacement = true;
+
                 if (!replacements.TryGetValue(current.Value.Span[indexOfOldChar], out ReadOnlyMemory<char> replacement))
                 {
                     replacement = ReadOnlyMemory<char>.Empty;
                 }
 
-                ReadOnlyMemory<char> value;
-                if (indexOfOldChar is 0)
+                if (indexOfOldChar > 0)
                 {
-                    value = replacement;
-                    replacementList = replacementList.Append(value.Span);
+                    replacementList = replacementList.Append(current.Value[..indexOfOldChar]);
                 }
-                else
-                {
-                    value = current.Value[..indexOfOldChar];
-                    replacementList = replacementList.Append(value.Span).Append(replacement.Span);
-                }
+
+                replacementList = replacementList.Append(replacement);
 
                 int index = indexOfOldChar;
                 foreach (int occurrence in occurrences.Skip(1))
@@ -396,34 +674,32 @@ public StringSegmentLinkedList Replace(Func<char, bool> predicate, IReadOnlyDict
 
                     if (index < occurrence)
                     {
-                        ReadOnlyMemory<char> valueToKeep = current.Value[( index + 1 ) .. occurrence];
-                        replacementList = replacementList.Append(valueToKeep.Span)
-                            .Append(replacement.Span);
+                        ReadOnlyMemory<char> valueToKeep = current.Value[(index + 1) .. occurrence];
+                        replacementList = replacementList.Append(valueToKeep)
+                            .Append(replacement);
                     }
                     else
                     {
-                        replacementList = replacementList.Append(replacement.Span);
+                        replacementList = replacementList.Append(replacement);
                     }
 
                     index = occurrence;
                 }
 
-                // we did all substitutions, but we did not reach the end of the original input
-                // => copy all remaining original chars starting at the index position
                 if (index < current.Value.Length)
                 {
-                    replacementList = replacementList.Append(current.Value[( index + 1 )..].Span);
+                    replacementList = replacementList.Append(current.Value[(index + 1)..]);
                 }
             }
             else
             {
-                replacementList = replacementList.Append(current.Value.Span);
+                replacementList = replacementList.Append(current.Value);
             }
 
             current = current.Next;
         }
 
-        return replacementList;
+        return anyReplacement ? replacementList : this;
     }
 
     /// <summary>
@@ -439,28 +715,38 @@ public StringSegmentLinkedList Replace(Func<char, bool> predicate, IReadOnlyDict
     /// </remarks>
     public StringSegmentLinkedList Replace(in ReadOnlySpan<char> oldValue, in ReadOnlySpan<char> newValue)
     {
-        StringSegmentLinkedList replacementList = [];
+        StringSegmentLinkedList replacementList = new();
         StringSegmentNode current = _head;
         int oldStringLength = oldValue.Length;
+        bool anyReplacement = false;
 
         while (current is not null)
         {
             ReadOnlySpan<char> subject = current.Value.Span;
-            int index = subject.IndexOf(oldValue[0]);
+            int index = subject.IndexOf(oldValue.Length > 0 ? oldValue[0] : '\0');
 
-            // no match in the current node
-            if (index is -1)
+            if (oldStringLength == 0)
             {
-                replacementList = replacementList.Append(subject);
+                // Nothing to replace
+                replacementList = replacementList.Append(current.Value);
+            }
+            else if (index is -1)
+            {
+                replacementList = replacementList.Append(current.Value);
             }
             else
             {
                 int previousIndex = 0;
                 do
                 {
-                    replacementList = replacementList.Append(subject[..index]);
+                    if (index > 0)
+                    {
+                        replacementList = replacementList.Append(subject[..index]);
+                    }
+
                     int offset = 0;
                     bool mismatchFound;
+                    int startIndex = index;
                     do
                     {
                         mismatchFound = subject[index] != oldValue[offset];
@@ -468,16 +754,19 @@ public StringSegmentLinkedList Replace(Func<char, bool> predicate, IReadOnlyDict
                         offset++;
                     } while (offset < oldStringLength && index < subject.Length && !mismatchFound);
 
-                    replacementList = mismatchFound switch
+                    if (mismatchFound)
                     {
-                        // a mismatch was found: we need to rewind and copy from the previous index up to the current one
-                        true => replacementList.Append(subject[(previousIndex + 1)..index]),
-                        // we looped through the whole span and, up to here, everything matches => we just append newValue
-                        _ => replacementList.Append(newValue)
-                    };
+                        // rewind and copy from previousIndex+1 to index
+                        replacementList = replacementList.Append(subject[(previousIndex + 1)..index]);
+                    }
+                    else
+                    {
+                        replacementList = replacementList.Append(newValue);
+                        anyReplacement = true;
+                    }
 
                     subject = subject[index..];
-                    previousIndex = index;
+                    previousIndex = startIndex;
                     index = subject.IndexOf(oldValue[0]);
                 } while (index != -1 && subject.Length > 0);
 
@@ -490,9 +779,16 @@ public StringSegmentLinkedList Replace(Func<char, bool> predicate, IReadOnlyDict
             current = current.Next;
         }
 
-        _head = replacementList._head;
+        if (anyReplacement)
+        {
+            _head = replacementList._head;
+            _tail = replacementList._tail;
+            Count = replacementList.Count;
+        }
 
-        return this;
+        return anyReplacement
+                   ? replacementList
+                   : this; // no-op when none replaced
     }
 
     /// <summary>
@@ -502,25 +798,48 @@ public StringSegmentLinkedList Replace(Func<char, bool> predicate, IReadOnlyDict
     /// <returns>A new <see cref="StringSegmentLinkedList"/> where all candidates <paramref name="predicate"/> were removed.</returns>
     public StringSegmentLinkedList RemoveBy(Func<ReadOnlyMemory<char>, bool> predicate)
     {
+        if (predicate is null)
+        {
+            throw new ArgumentNullException(nameof(predicate));
+        }
+
         StringSegmentNode current = _head;
         StringSegmentNode previous = null;
+        int newCount = 0;
+
         while (current is not null)
         {
             if (predicate(current.Value))
             {
+                // remove current
                 if (previous is not null)
                 {
                     previous.Next = current.Next;
                 }
                 else
                 {
-                    Debug.Assert(_head.Equals(current));
-                    _head = _tail;
+                    _head = current.Next;
+                }
+
+                if (current.Next is null)
+                {
+                    _tail = previous;
                 }
             }
+            else
+            {
+                // keep current
+                previous = current;
+                newCount++;
+            }
 
-            previous = current;
             current = current.Next;
+        }
+
+        Count = newCount;
+        if (Count is 0)
+        {
+            _head = _tail = null;
         }
 
         return this;
@@ -539,31 +858,19 @@ public StringSegmentLinkedList Replace(Func<char, bool> predicate, IReadOnlyDict
             throw new ArgumentNullException(nameof(other));
         }
 
-        StringSegmentLinkedList result = [];
+        StringSegmentLinkedList result = new();
         StringSegmentNode current = _head;
-        if (!current.Equals(EmptyNode))
+        while (current is not null)
         {
-            while (current is not null)
-            {
-                result = result.Append(current.Value.Span);
-                current = current.Next;
-            }
+            result = result.Append(current.Value);
+            current = current.Next;
         }
 
         current = other._head;
-        if (!current.Equals(EmptyNode))
+        while (current is not null)
         {
-            if (result.Count is 0)
-            {
-                result = new StringSegmentLinkedList(current.Value.Span);
-                current = current.Next;
-            }
-
-            while (current is not null)
-            {
-                result = result.Append(current.Value.Span);
-                current = current.Next;
-            }
+            result = result.Append(current.Value);
+            current = current.Next;
         }
 
         return result;
@@ -761,7 +1068,7 @@ public StringSegmentLinkedList Replace(Func<char, bool> predicate, IReadOnlyDict
             int searchIndex = 0;
             while (currentNode is not null && !found)
             {
-                var valueSpan = currentNode.Value.Span;
+                ReadOnlySpan<char> valueSpan = currentNode.Value.Span;
 
                 for (int i = 0; i < valueSpan.Length && !found; i++)
                 {
@@ -809,7 +1116,7 @@ public StringSegmentLinkedList Replace(Func<char, bool> predicate, IReadOnlyDict
 
         if (search.IsEmpty)
         {
-            startsWith = EmptyNode.Equals(_head);
+            startsWith = true;
         }
         else
         {
@@ -850,6 +1157,7 @@ public StringSegmentLinkedList Replace(Func<char, bool> predicate, IReadOnlyDict
                     {
                         offset += current.Value.Length;
                     }
+
                     current = current.Next;
                 } while (offset <= search.Length && !mismatchFound && current is not null);